@@ -1,20 +1,17 @@
+'use client'
+
 import React from 'react';
-import { useRouter } from 'next/router';
-import { LayoutTemplate, SEO, SEOProps } from '@bbl-nx/ui-components';
+import { LayoutTemplate } from '@bbl-nx/ui-components/templates/layout-template/layout-template';
 
 interface LayoutProps {
-  seoProps?: Omit<SEOProps, 'asPath'>;
   children: React.ReactNode;
 }
 
 export default function Layout(props: LayoutProps) {
-  const { seoProps, children } = props;
-  const router = useRouter();
-  const { asPath } = router;
+  const { children } = props;
 
   return (
     <>
-      <SEO asPath={asPath} {...seoProps} />
       <LayoutTemplate>{children}</LayoutTemplate>
     </>
   );
