@@ -1,5 +1,5 @@
 import { home } from '@bbl-nx/constants';
-import { HomeCard } from '../../organisms';
+import { HomeCard } from '../../organisms/home-card/home-card';
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface HomeTemplateProps {}
